added param to bag_recorder

+++ bag_recorder/bag_recorder/bag_recorder.py
@@ -15,7 +15,8 @@ import rclpy
 import subprocess
 from rclpy.lifecycle import LifecycleNode, TransitionCallbackReturn
 
-
+# Import service
+from project_interfaces.srv import SaveSrv
 
 
 class BagRecorder(LifecycleNode):
@@ -28,6 +29,10 @@ class BagRecorder(LifecycleNode):
 		self.declare_parameter('topics', ['/raw_image_out'])
 		self.declare_parameter('bag_name', 'test_bag')
 		self.declare_parameter('save_directory', '.')
+		self.declare_parameter('is_saving', False)
+
+		self.bag_process = None
+
 
 	def on_configure(self, previous_state):
 		
@@ -37,6 +42,7 @@ class BagRecorder(LifecycleNode):
 		self.topics = self.get_parameter('topics').value
 		self.bag_name = self.get_parameter('bag_name').value
 		self.save_directory = self.get_parameter('save_directory').value
+		self.is_saving = self.get_parameter('is_saving').value
 
 		# Establish filepaths
 		os.makedirs(self.save_directory, exist_ok=True)
@@ -44,26 +50,32 @@ class BagRecorder(LifecycleNode):
 		# Counter to make sure we don't override bags if paused.
 		self.counter = 1
 
+
 		return TransitionCallbackReturn.SUCCESS
 
-
-
 	def on_activate(self, previous_state):
-		self.get_logger().info('Activating ROS2 bag recording')
 
-		file_name = f'{self.bag_name}_{self.counter}'
+		if self.is_saving:
 
-		self.bag_path = os.path.join(self.save_directory, file_name)
+			self.get_logger().info('Activating ROS2 bag recording')
 
-		# Create the command and use Popen to run it
-		cmd = ['ros2', 'bag', 'record', '-o', self.bag_path] + self.topics
-		self.get_logger().info(f"Running: {' '.join(cmd)}")
+			file_name = f'{self.bag_name}_{self.counter}'
 
-		try:
-			self.bag_process = subprocess.Popen(cmd)
-		except Exception as e:
-			self.get_logger().error(f"Failed to start bag recording :{e}")
-			return TransitionCallbackReturn.FAILURE
+			self.bag_path = os.path.join(self.save_directory, file_name)
+
+			# Create the command and use Popen to run it
+			cmd = ['ros2', 'bag', 'record', '-o', self.bag_path] + self.topics
+			self.get_logger().info(f"Running: {' '.join(cmd)}")
+
+			try:
+				self.bag_process = subprocess.Popen(cmd)
+			except Exception as e:
+				self.get_logger().error(f"Failed to start bag recording :{e}")
+				return TransitionCallbackReturn.FAILURE
+			
+		else:
+			self.get_logger().info('Bag recording skipped (is_saving set to False)')
+
 	
 		return super().on_activate(previous_state)
 	
@@ -71,13 +83,15 @@ class BagRecorder(LifecycleNode):
 	def on_deactivate(self, previous_state):
 		self.get_logger().info('Deactivating ROS2 bag recording')
 
-		if self.bag_process:
-			self.bag_process.terminate()
-			self.bag_process.wait()
-			self.bag_process = None
+		if self.is_saving:
 
-		# Increment counter for next recording name
-		self.counter += 1
+			if self.bag_process:
+				self.bag_process.terminate()
+				self.bag_process.wait()
+				self.bag_process = None
+
+			# Increment counter for next recording name
+			self.counter += 1
 
 		return super().on_deactivate(previous_state)
 	
@@ -85,15 +99,23 @@ class BagRecorder(LifecycleNode):
 	def on_cleanup(self, previous_state):
 		self.get_logger().info('Cleaning up')
 
+		if self.is_saving:
+
+			if self.bag_process:
+				self.bag_process.terminate()
+				self.bag_process.wait()
+
 		return TransitionCallbackReturn.SUCCESS
 
 	
 	def on_shutdown(self, previous_state):
 		self.get_logger().info('Shutting down')
 		
-		if self.bag_process:
-			self.bag_process.terminate()
-			self.bag_process.wait()
+		if self.is_saving:
+
+			if self.bag_process:
+				self.bag_process.terminate()
+				self.bag_process.wait()
 
 		return TransitionCallbackReturn.SUCCESS
 
@@ -102,7 +124,6 @@ class BagRecorder(LifecycleNode):
 		self.get_logger().error('ERROR!')
 
 		return TransitionCallbackReturn.ERROR
-
 
 
 def main(args=None):
